Do CLI import in __main__, not __init__
Importing the CLI code doubles the import time for httpx altogether. By
importing in httpx/__main__.py instead of in httpx/__init__.py, the
additional time only happens when running the httpx CLI, and not on
every "import httpx".
