+++ httpx/__init__.py
@@ -11,20 +11,6 @@ from ._transports import *
 from ._types import *
 from ._urls import *
 
-try:
-    from ._main import main
-except ImportError:  # pragma: no cover
-
-    def main() -> None:  # type: ignore
-        import sys
-
-        print(
-            "The httpx command line client could not run because the required "
-            "dependencies were not installed.\nMake sure you've installed "
-            "everything with: pip install 'httpx[cli]'"
-        )
-        sys.exit(1)
-
 
 __all__ = [
     "__description__",
@@ -59,7 +45,6 @@ __all__ = [
     "InvalidURL",
     "Limits",
     "LocalProtocolError",
-    "main",
     "MockTransport",
     "NetRCAuth",
     "NetworkError",
